[fix/feat] fix ANTs installation
The ANTs installation would fail with a 403 not found error because the
archive is no longer available on Dropbox. I re-created the archive from
an old iCVMapp3r image and migrated it to the AICONs Lab HuggingFace
account as a dataset. I changed/improved the cmds in the Dockerfile
accordingly and also fixed deprecated syntaxt to set env vars elsewhere
in the Dockerfile.